Sensible error message on missing template file

+++ juxta.sh
@@ -205,6 +205,10 @@ log2() {
 # http://stackoverflow.com/questions/14434549/how-to-expand-shell-variables-in-a-text-file
 # Input: template-file
 function ctemplate() {
+    if [[ ! -s "$1" ]]; then
+        >&2 echo "Error: Template '$1' could not be found"
+        exit 8
+    fi
     local TMP=$(mktemp /tmp/juxta_XXXXXXXX)
     echo 'cat <<END_OF_TEXT' >  "$TMP"
     cat  "$1"                >> "$TMP"
